requirements.txt
aliyun-python-sdk-core-v3==2.13.11
amqp==2.5.2
appdirs==1.4.4
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
Automat==20.2.0
Babel==2.8.0
baron==0.3
beautifulsoup4==4.9.1
billiard==3.6.2.0
boto3==1.12.6
botocore==1.15.6
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cssselect==1.1.0
Django==2.1.13
django-appconf==1.0.3
django-celery-results==1.2.0
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-filter==2.2.0
django-imagekit==4.0.2
django-redis==4.11.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
dnspython==1.16.0
docutils==0.15.2
elasticsearch==7.5.1
elasticsearch-dsl==7.2.1
et-xmlfile==1.0.1
eventlet==0.25.1
fake-useragent==0.1.11
Faker==5.0.1
flower==0.9.3
gevent==1.4.0
greenlet==0.4.15
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.5.0
incremental==17.5.0
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4.1
jieba==0.42.1
Jinja2==2.11.1
jmespath==0.9.5
kombu==4.6.7
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
more-itertools==8.2.0
mysqlclient==1.4.6
numpy==1.18.1
openapi-codec==1.3.2
openpyxl==3.0.3
packaging==20.1
pandas==1.0.2
parsel==1.5.2
pathlib==1.0.1
pilkit==2.0
Pillow==7.0.0
pluggy==0.13.1
Protego==0.1.16
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.4.4
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
pypinyin==0.37.0
pytest==5.3.5
python-dateutil==2.8.1
pytz==2019.3
qiniu==7.2.6
qrcode==6.1
queuelib==1.5.0
redis==3.4.1
requests==2.23.0
rply==0.7.7
s3transfer==0.3.3
Scrapy==2.0.0
service-identity==18.1.0
simplejson==3.17.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.0
text-unidecode==1.3
tornado==5.1.1
Twisted==19.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
w3lib==1.21.0
wcwidth==0.1.8
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
zipp==3.0.0
zope.interface==4.7.1
